requirements.txt
annotated-types==0.7.0
anthropic==0.52.1
anyio==4.9.0
blockbuster==1.5.24
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
cryptography==44.0.3
distro==1.9.0
forbiddenfruit==0.1.4
greenlet==3.2.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
isort==6.0.1
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema_rs==0.29.1
langchain==0.3.25
langchain-anthropic==0.3.14
langchain-core==0.3.62
langchain-ollama==0.3.3
langchain-text-splitters==0.3.8
langgraph==0.4.7
langgraph-api==0.2.34
langgraph-checkpoint==2.0.26
langgraph-cli==0.2.10
langgraph-prebuilt==0.2.2
langgraph-runtime-inmem==0.2.0
langgraph-sdk==0.1.70
langsmith==0.3.42
ollama==0.4.9
orjson==3.10.18
ormsgpack==1.10.0
packaging==24.2
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
PyJWT==2.10.1
python-dotenv==1.1.0
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
ruff==0.11.12
sniffio==1.3.1
SQLAlchemy==2.0.41
sse-starlette==2.1.3
starlette==0.46.2
structlog==25.3.0
tenacity==9.1.2
truststore==0.10.1
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
watchfiles==1.0.5
xxhash==3.5.0
zstandard==0.23.0
